Cleaner tmpdir handling and setting `CREATE_UNIQUE_DIR` as `True` by default (#1578)
1. `CREATE_UNIQUE_DIR` is now set to `True` by default (closes #1477)
2. The tmpdir for actively running jobs will now have the same timestamp
as the unique result folder where the data is stored. Before, they
differed slightly, which did not make much sense.
2. The tmpdir is named `tmp-quacc-...` instead of `quacc-tmp...` for
easier indexing/deleting. Similarly, the "symlink" now prepends the name
instead of being placed at the end.

---------

+++ src/quacc/runners/prep.py
@@ -2,10 +2,8 @@
 from __future__ import annotations
 
 import os
-from datetime import datetime, timezone
 from pathlib import Path
 from shutil import move, rmtree
-from tempfile import mkdtemp
 
 from monty.shutil import gzip_dir, remove
 
@@ -33,33 +31,30 @@ def calc_setup(
     Returns
     -------
     Path
-        The path to the tmpdir, where the calculation will be run. It will be
+        The path to the unique tmpdir, where the calculation will be run. It will be
         deleted after the calculation is complete. By default, this will be
         located within the `SETTINGS.SCRATCH_DIR`, but if that is not set, it will
-        be located within the `SETTINGS.RESULTS_DIR`.
+        be located within the `SETTINGS.RESULTS_DIR`. For conenience, a symlink
+        to this directory will be made in the `SETTINGS.RESULTS_DIR`.
     Path
         The path to the results_dir, where the files will ultimately be stored.
-        A symlink to the tmpdir will be made here during the calculation for
-        convenience. By defualt, this will be the `SETTINGS.RESULTS_DIR`, but if
+        By defualt, this will be the `SETTINGS.RESULTS_DIR`, but if
         `SETTINGS.CREATE_UNIQUE_DIR` is set, it will be a unique directory
         within the `SETTINGS.RESULTS_DIR`.
     """
 
-    # Set where to store the results
-    job_results_dir = (
-        make_unique_dir(base_path=SETTINGS.RESULTS_DIR)
-        if SETTINGS.CREATE_UNIQUE_DIR
-        else SETTINGS.RESULTS_DIR
-    )
-    tmpdir_base = SETTINGS.SCRATCH_DIR or SETTINGS.RESULTS_DIR
-
     # Create a tmpdir for the calculation
-    time_now = datetime.now(timezone.utc).strftime("%Y-%m-%d-%H-%M-%S-%f")
-    tmpdir = Path(mkdtemp(prefix=f"quacc-tmp-{time_now}-", dir=tmpdir_base)).resolve()
+    tmpdir_base = SETTINGS.SCRATCH_DIR or SETTINGS.RESULTS_DIR
+    tmpdir = make_unique_dir(base_path=tmpdir_base, prefix="tmp-quacc-")
 
-    # Create a symlink to the tmpdir in the results_dir
+    # Define the results directory
+    job_results_dir = SETTINGS.RESULTS_DIR
+    if SETTINGS.CREATE_UNIQUE_DIR:
+        job_results_dir /= f"{tmpdir.name.split('tmp-')[-1]}"
+
+    # Create a symlink to the tmpdir
     if os.name != "nt" and SETTINGS.SCRATCH_DIR:
-        symlink = job_results_dir / f"{tmpdir.name}-symlink"
+        symlink = SETTINGS.RESULTS_DIR / f"symlink-{tmpdir.name}"
         symlink.unlink(missing_ok=True)
         symlink.symlink_to(tmpdir, target_is_directory=True)
 
@@ -96,7 +91,8 @@ def calc_cleanup(tmpdir: Path, job_results_dir: Path) -> None:
     None
     """
 
-    # Change to the results directory
+    # Make the results directory and cd into it
+    Path(job_results_dir).mkdir(parents=True, exist_ok=True)
     os.chdir(job_results_dir)
 
     # Gzip files in tmpdir
@@ -110,8 +106,9 @@ def calc_cleanup(tmpdir: Path, job_results_dir: Path) -> None:
         move(tmpdir / file_name, job_results_dir / file_name)
 
     # Remove symlink to tmpdir
-    symlink_path = job_results_dir / f"{tmpdir.name}-symlink"
-    symlink_path.unlink(missing_ok=True)
+    if os.name != "nt" and SETTINGS.SCRATCH_DIR:
+        symlink_path = SETTINGS.RESULTS_DIR / f"symlink-{tmpdir.name}"
+        symlink_path.unlink(missing_ok=True)
 
     # Remove the tmpdir
     rmtree(tmpdir, ignore_errors=True)
